+++ udata/__init__.py
@@ -4,5 +4,5 @@
 udata
 """
 
-__version__ = "10.6.1.dev"
+__version__ = "10.8.3.dev"
 __description__ = "Open data portal"