+++ udata/harvest/tests/test_dcat_backend.py
@@ -74,7 +74,7 @@ class DcatBackendTest:
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
 
-        actions.run(source.slug)
+        actions.run(source)
 
         source.reload()
 
@@ -123,7 +123,7 @@ class DcatBackendTest:
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
 
-        actions.run(source.slug)
+        actions.run(source)
 
         datasets = {d.harvest.dct_identifier: d for d in Dataset.objects}
 
@@ -154,7 +154,7 @@ class DcatBackendTest:
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
 
-        actions.run(source.slug)
+        actions.run(source)
 
         datasets = {d.harvest.dct_identifier: d for d in Dataset.objects}
 
@@ -171,7 +171,7 @@ class DcatBackendTest:
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
 
-        actions.run(source.slug)
+        actions.run(source)
 
         dataservices = Dataservice.objects
 
@@ -187,12 +187,29 @@ class DcatBackendTest:
             == "https://data.paris2024.org/api/explore/v2.1/console"
         )
 
+    def test_harvest_dataservices_ignore_accessservices(self, rmock):
+        rmock.get("https://example.com/schemas", json=ResourceSchemaMockData.get_mock_data())
+
+        url = mock_dcat(rmock, "catalog.xml")
+        org = OrganizationFactory()
+        source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
+
+        actions.run(source)
+
+        source.reload()
+
+        job = source.get_last_job()
+        assert len(job.items) == 4
+
+        dataservices = Dataservice.objects
+        assert len(dataservices) == 0
+
     def test_harvest_literal_spatial(self, rmock):
         url = mock_dcat(rmock, "evian.json")
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
 
-        actions.run(source.slug)
+        actions.run(source)
 
         datasets = {d.harvest.dct_identifier: d for d in Dataset.objects}
         assert len(datasets) == 8
@@ -244,7 +261,7 @@ class DcatBackendTest:
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
 
-        actions.run(source.slug)
+        actions.run(source)
 
         datasets = {d.harvest.dct_identifier: d for d in Dataset.objects}
 
@@ -306,7 +323,7 @@ class DcatBackendTest:
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
 
-        actions.run(source.slug)
+        actions.run(source)
 
         assert Dataset.objects.count() == 2
         assert HarvestJob.objects.first().status == "done"
@@ -320,7 +337,7 @@ class DcatBackendTest:
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
 
-        actions.run(source.slug)
+        actions.run(source)
 
         datasets = {d.harvest.dct_identifier: d for d in Dataset.objects}
 
@@ -345,7 +362,7 @@ class DcatBackendTest:
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
 
-        actions.run(source.slug)
+        actions.run(source)
 
         datasets = {d.harvest.dct_identifier: d for d in Dataset.objects}
 
@@ -388,7 +405,7 @@ class DcatBackendTest:
         url = mock_dcat(rmock, filename)
         source = HarvestSourceFactory(backend="dcat", url=url, organization=OrganizationFactory())
 
-        actions.run(source.slug)
+        actions.run(source)
 
         source.reload()
 
@@ -416,8 +433,8 @@ class DcatBackendTest:
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
 
         # Run the same havester twice
-        actions.run(source.slug)
-        actions.run(source.slug)
+        actions.run(source)
+        actions.run(source)
 
         datasets = {d.harvest.dct_identifier: d for d in Dataset.objects}
 
@@ -431,7 +448,7 @@ class DcatBackendTest:
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
 
-        actions.run(source.slug)
+        actions.run(source)
 
         source.reload()
 
@@ -443,7 +460,7 @@ class DcatBackendTest:
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
 
-        actions.run(source.slug)
+        actions.run(source)
 
         source.reload()
 
@@ -456,7 +473,7 @@ class DcatBackendTest:
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
 
-        actions.run(source.slug)
+        actions.run(source)
 
         source.reload()
 
@@ -470,7 +487,7 @@ class DcatBackendTest:
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
 
-        actions.run(source.slug)
+        actions.run(source)
 
         source.reload()
 
@@ -478,12 +495,8 @@ class DcatBackendTest:
 
         assert job.status == "done"
         assert job.errors == []
-        assert len(job.items) == 5
-        # 4 datasets and one Dataservice mentionned but not described
-        # because it appears in a distribution as DCAT.accessService
-        # but is missing a proper DCT.identifier
+        assert len(job.items) == 4
         assert len([item for item in job.items if item.status == "done"]) == 4
-        assert len([item for item in job.items if item.status == "skipped"]) == 1
 
     def test_xml_catalog(self, rmock):
         LicenseFactory(id="lov2", title="Licence Ouverte Version 2.0")
@@ -493,7 +506,7 @@ class DcatBackendTest:
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
 
-        actions.run(source.slug)
+        actions.run(source)
 
         # test dct:license support
         dataset = Dataset.objects.get(harvest__dct_identifier="3")
@@ -587,7 +600,7 @@ class DcatBackendTest:
         url = mock_dcat(rmock, "geonetwork.xml", path="catalog.xml")
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
-        actions.run(source.slug)
+        actions.run(source)
         dataset = Dataset.objects.filter(organization=org).first()
         assert dataset is not None
         assert dataset.harvest is not None
@@ -616,7 +629,7 @@ class DcatBackendTest:
         url = mock_dcat(rmock, "sig.oreme.rdf")
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
-        actions.run(source.slug)
+        actions.run(source)
         dataset = Dataset.objects.filter(organization=org).first()
 
         assert dataset is not None
@@ -646,7 +659,7 @@ class DcatBackendTest:
         url = mock_dcat(rmock, "udata.xml")
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
-        actions.run(source.slug)
+        actions.run(source)
 
         source.reload()
         job = source.get_last_job()
@@ -712,7 +725,7 @@ class DcatBackendTest:
         get_mock = rmock.get(url)
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
-        actions.run(source.slug)
+        actions.run(source)
 
         assert "User-Agent" in get_mock.last_request.headers
         assert get_mock.last_request.headers["User-Agent"] == "uData/0.1 dcat"
@@ -723,7 +736,7 @@ class DcatBackendTest:
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
 
-        actions.run(source.slug)
+        actions.run(source)
 
         source.reload()
 
@@ -741,7 +754,7 @@ class DcatBackendTest:
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
 
-        actions.run(source.slug)
+        actions.run(source)
 
         source.reload()
 
@@ -773,7 +786,7 @@ class DcatBackendTest:
         rmock.head(url, headers={"Content-Type": "application/json"})
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="dcat", url=url, organization=org)
-        actions.run(source.slug)
+        actions.run(source)
 
         source.reload()
 
@@ -787,7 +800,7 @@ class DcatBackendTest:
         rmock.get(url, status_code=404)
 
         source = HarvestSourceFactory(backend="dcat", url=url, organization=OrganizationFactory())
-        actions.run(source.slug)
+        actions.run(source)
         source.reload()
 
         job = source.get_last_job()
@@ -800,7 +813,7 @@ class DcatBackendTest:
         rmock.head(url, status_code=404)
 
         source = HarvestSourceFactory(backend="dcat", url=url, organization=OrganizationFactory())
-        actions.run(source.slug)
+        actions.run(source)
         source.reload()
 
         job = source.get_last_job()
@@ -817,7 +830,7 @@ class CswDcatBackendTest:
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="csw-dcat", url=url, organization=org)
 
-        actions.run(source.slug)
+        actions.run(source)
 
         source.reload()
 
@@ -863,7 +876,7 @@ class CswDcatBackendTest:
         org = OrganizationFactory()
         source = HarvestSourceFactory(backend="csw-dcat", url=url, organization=org)
 
-        actions.run(source.slug)
+        actions.run(source)
 
         assert "User-Agent" in get_mock.last_request.headers
         assert get_mock.last_request.headers["User-Agent"] == "uData/0.1 csw-dcat"
@@ -886,7 +899,7 @@ class CswIso19139DcatBackendTest:
         with open(os.path.join(CSW_DCAT_FILES_DIR, "XSLT.xml"), "r") as f:
             xslt = f.read()
         url = mock_csw_pagination(rmock, "geonetwork/srv/eng/csw.rdf", "geonetwork-iso-page-{}.xml")
-        rmock.get(current_app.config.get("HARVEST_ISO19139_XSL_URL"), text=xslt)
+        rmock.get(current_app.config.get("HARVEST_ISO19139_XSLT_URL"), text=xslt)
         org = OrganizationFactory()
         source = HarvestSourceFactory(
             backend="csw-iso-19139",
@@ -902,7 +915,7 @@ class CswIso19139DcatBackendTest:
             },
         )
 
-        actions.run(source.slug)
+        actions.run(source)
 
         source.reload()