+++ CHANGELOG.md
@@ -2,7 +2,48 @@
 
 ## Current (in progress)
 
+- Count dataservices and reuses for datasets based on signal [#3335](https://github.com/opendatateam/udata/pull/3335)
+  :warning: the job `update-datasets-reuses-metrics` disappears, you should unschedule it before installing this version
 - Add support for all GeoDCAT-AP roles [#3374](https://github.com/opendatateam/udata/pull/3374)
+- Consolidate CSW harvesting logic [#3378](https://github.com/opendatateam/udata/pull/3378)
+  :warning: Config setting `HARVEST_ISO19139_XSL_URL` is renamed to `HARVEST_ISO19139_XSLT_URL`.
+- Compute organization aggregate metrics only once at the end of harvest [#3391](https://github.com/opendatateam/udata/pull/3391)
+- Cache dcat harvest pages to avoid two rounds of requests [#3398](https://github.com/opendatateam/udata/pull/3398)
+- Ignore Dataset.accessService when processing DataService [#3399](https://github.com/opendatateam/udata/pull/3399)
+- Add dataset field `description_short` [#3397](https://github.com/opendatateam/udata/pull/3397)
+
+## 10.8.2 (2025-07-31)
+
+- Add integrity checks on non existent list field instead of empty [#3242](https://github.com/opendatateam/udata/pull/3242)
+- Migrate udata-metrics [#3386](https://github.com/opendatateam/udata/pull/3386/)
+- Migrate udata-ckan [#3385](https://github.com/opendatateam/udata/pull/3385)
+- Deactivate sitemap extension [#3389](https://github.com/opendatateam/udata/pull/3389)
+- Deactivate cache to check for existing slugs [#3390](https://github.com/opendatateam/udata/pull/3390)
+- Fix not returning 404 for harvest sources [#3382](https://github.com/opendatateam/udata/pull/3382/)
+- Add missing trailing slashes on some endpoints [#3392](https://github.com/opendatateam/udata/pull/3392/)
+- Make organization name filter case insensitive [#3395](https://github.com/opendatateam/udata/pull/3395)
+- Harvest metadata documents are no longer dynamic [#3396](https://github.com/opendatateam/udata/pull/3396)
+
+## 10.8.1 (2025-07-25)
+
+- Fix user deletion performances issues [#3388](https://github.com/opendatateam/udata/pull/3388)
+- Add search to discussions [#3384](https://github.com/opendatateam/udata/pull/3384)
+- Separate metrics for opened discussions [#3370](https://github.com/opendatateam/udata/pull/3370)
+
+## 10.8.0 (2025-07-22)
+
+- Remove placeholder Jinja filter [#3372](https://github.com/opendatateam/udata/pull/3372/)
+- Add `self_web_url` for discussion [#3375](https://github.com/opendatateam/udata/pull/3375)
+- Fix dataset schemas endpoint when no resource [#3373](https://github.com/opendatateam/udata/pull/3373)
+- Add last_update as a dataset model property [#3377](https://github.com/opendatateam/udata/pull/3377)
+- Add `label` (translation of role) to API response for org members [#3383](https://github.com/opendatateam/udata/pull/3383)
+- Fix crashing on datasets with no slugs [#3368](https://github.com/opendatateam/udata/pull/3368/)
+- Fix duplicate contact point creation [#3379](https://github.com/opendatateam/udata/pull/3379) [#3381](https://github.com/opendatateam/udata/pull/3381)
+- Atomic `get_or_create` [#3366](https://github.com/opendatateam/udata/pull/3366)
+- Add contact points suggest [#3376](https://github.com/opendatateam/udata/pull/3376)
+
+## 10.7.0 (2025-07-17)
+
 - Add name and siret filters for orgs [#3362](https://github.com/opendatateam/udata/pull/3362)
 - Refactor how to get URLs to the cdata front-end, need to set the `CDATA_BASE_URL` env variable (removal of `endpoint_for()`, `fields.UrlFor()`, `display_url`, `external_url` and `external_url_with_campaign`) [#3345](https://github.com/opendatateam/udata/pull/3345) [#3371](https://github.com/opendatateam/udata/pull/3371)
 - Add back preview_url from tabular API (remove the PreviewPlugin system) [#3364](https://github.com/opendatateam/udata/pull/3364)